+++ src/app/index.run.js
@@ -1,3 +1,4 @@
+/*eslint angular/no-private-call: [2,{"allow":["$$route"]}]*/
 (function() {
   'use strict';
 
@@ -25,5 +26,6 @@
                 });
             }
         });
+        $rootScope.$on('$destroy', routeChange);
     }
 })();